% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{as.matrix.fcwtr_scalogram}
\alias{as.matrix.fcwtr_scalogram}
\title{Extract the data matrix from a scalogram}
\usage{
\method{as.matrix}{fcwtr_scalogram}(x, ...)
}
\arguments{
\item{x}{An "fcwtr_scalogram" object resulting from \code{\link[=fcwt]{fcwt()}}.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
A two dimensional numeric vector, inheriting the class "matrix".
}
\description{
Strips attributes and class from a scalogram object to retrieve
a pure matrix.
}
\examples{
options(max.print = 10)

fcwt(
  sin((1:5000) * 2 * pi * 440 / 44100),
  x_sample_freq = u(44100, "Hz"),
  n_freqs = 10
) |>
  as.matrix()

}
