% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{as_tibble.fcwtr_scalogram}
\alias{as_tibble.fcwtr_scalogram}
\title{Convert the scalogram matrix to a tibble}
\usage{
\method{as_tibble}{fcwtr_scalogram}(x, ...)
}
\arguments{
\item{x}{An "fcwtr_scalogram" object resulting from \code{\link[=fcwt]{fcwt()}}.}

\item{...}{Additional arguments are ignored.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} object representing the scalogram data with four columns:
\describe{
\item{time_index}{An integer index uniquely identifying time slices.}
\item{time}{The time difference to the first time slice in physical units.
The time unit is the inverse of the frequency unit chosen by the user
for the \code{sample_freq} argument of \code{\link[=fcwt]{fcwt()}}.}
\item{freq}{The frequency in the same units as the \code{sample_freq} argument
of \code{\link[=fcwt]{fcwt()}}.}
\item{value}{The fCWT result for the particular time-frequency combination.}
}
}
\description{
Internally, the scalogram resulting from \code{\link[=fcwt]{fcwt()}} is represented by
a numeric \code{\link[=matrix]{matrix()}}. This method converts this matrix into a three-columned
\code{\link[tibble:tibble]{tibble::tibble()}} (two columns to represent the matrix indices, and one column
for the corresponding value).
Note that this conversion has a significant run time cost.
}
\details{
This method is conceptually identical to \code{\link[=as.data.frame]{as.data.frame()}},
it simply returns a \code{\link[tibble:tibble]{tibble::tibble()}} instead of a \code{\link[=data.frame]{data.frame()}}.
}
\examples{
\dontshow{if (requireNamespace("tibble", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(tibble)

fcwt(
  sin((1:5000) * 2 * pi * 440 / 44100),
  x_sample_freq = u(44100, "Hz"),
  n_freqs = 10
) |>
  as_tibble()
\dontshow{\}) # examplesIf}
}
