% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{autoplot.fcwtr_scalogram}
\alias{autoplot.fcwtr_scalogram}
\title{Create a ggplot object resembling a scalogram}
\usage{
autoplot.fcwtr_scalogram(object, n = 1000, ...)
}
\arguments{
\item{n}{The plotting function reduces the time resolution by averaging
to generate a reasonable graphics format. \code{n} is the number of time
steps that are plotted. Defaults to \code{n = 1000}.}

\item{...}{other arguments passed to specific methods}
}
\value{
A ggplot object.
}
\description{
Create a ggplot object resembling a scalogram
}
\keyword{internal}
