% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_support.R
\name{avx_enabled}
\alias{avx_enabled}
\title{Check for AVX instruction set support}
\usage{
avx_enabled()
}
\value{
( \code{TRUE} | \code{FALSE} )
Returns \code{TRUE} if AVX support is enabled, \code{FALSE} otherwise.
}
\description{
This function checks if fCWTr was compiled with AVX instruction set support.
AVX instructions need to be supported by the users' CPU in order for this
to work.
}
\details{
By default, most compiler setups do not make use of AVX to increase
portability of the binary. If you are an R user that has a CPU supporting
AVX and want to make use of it, you might need to manually enable compiler
flags to let R know about it, and install the package from source
(so that it gets compiled on your machine).
}
\examples{
avx_enabled()

}
\concept{sys_support}
