% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_schemes.R
\name{create_opt_schemes}
\alias{create_opt_schemes}
\title{Create Optimization Schemes}
\usage{
create_opt_schemes(max_size, nthreads = 8L, flag = "measure")
}
\arguments{
\item{max_size}{Typical length of input vector}

\item{nthreads}{Number of threads used in the computation}

\item{flag}{From https://www.fftw.org/fftw3_doc/Planner-Flags.html:
"estimate"" specifies that, instead of actual measurements of different algorithms, a simple heuristic is used to pick a (probably sub-optimal) plan quickly. With this flag, the input/output arrays are not overwritten during planning.
"measure"" tells FFTW to find an optimized plan by actually computing several FFTs and measuring their execution time. Depending on your machine, this can take some time (often a few seconds). FFTW_MEASURE is the default planning option.
"patient"" is like FFTW_MEASURE, but considers a wider range of algorithms and often produces a “more optimal” plan (especially for large transforms), but at the expense of several times longer planning time (especially for large transforms).
"exhaustive" is like FFTW_PATIENT, but considers an even wider range of algorithms, including many that we think are unlikely to be fast, to produce the most optimal plan but with a substantially increased planning time.
"wisdom_only" is a special planning mode in which the plan is only created if wisdom is available for the given problem, and otherwise a NULL plan is returned. This can be combined with other flags, e.g. ‘FFTW_WISDOM_ONLY | FFTW_PATIENT’ creates a plan only if wisdom is available that was created in FFTW_PATIENT or FFTW_EXHAUSTIVE mode. The FFTW_WISDOM_ONLY flag is intended for users who need to detect whether wisdom is available; for example, if wisdom is not available one may wish to allocate new arrays for planning so that user data is not overwritten.}
}
\description{
Utilizes an FFTW feature to create optimization schemes
before the actual calulations. In conjunction with 'fcwt(...,optplan = T)`
this can provide a performance boost.
}
\seealso{
fcwt
}
