% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{du}
\alias{du}
\title{Drop physical units from a quantity}
\usage{
du(x, unit)
}
\arguments{
\item{x}{A dimensionful numeric vector (S3 class "units"), e.g. generated with \code{\link[=u]{u()}} or
subsequent calculations.}

\item{unit}{A character indicating the unit. See \code{\link[units:valid_udunits]{units::valid_udunits()}} and
\code{\link[units:valid_udunits]{units::valid_udunits_prefixes()}} for possible units.}
}
\value{
A dimensionless number defined by the ratio of \code{x} and one unit \code{u(1, unit)}.
}
\description{
Drop units from a dimensionful quantity and retrieve a dimensionless number.
The dimensionless number is defined by the ratio of \code{x} and one unit \code{u(1, unit)}.
}
\examples{
 x <- u(1, "km")

 # retrieve a dimensionless number
 # from x in units of "meter"
 du(x, "m")

}
\seealso{
\code{\link[units:drop_units]{units::drop_units()}}
}
