% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwt.R
\name{fcwt}
\alias{fcwt}
\title{Performs a fast continuous wavelet transform}
\usage{
fcwt(
  input,
  sample_freq,
  freq_begin,
  freq_end,
  n_freqs,
  sigma,
  abs = FALSE,
  nthreads = 8L
)
}
\arguments{
\item{noctave}{The number of octaves to be processed.}

\item{optplan}{If TRUE uses the optimization plans that might have been
generated before by [create_optimization_schemes_raw].}
}
\value{
Returns a numeric vector with dim = c(2, length(input), scales)
         where the first slot represents the real and imaginary part of the
         complex result, the second slot represents the time index, and
         the third slot represent different scales / frequencies.
}
\description{
This function call is a thin wrapper for the fcwt API.
}
