% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fcwt_raw}
\alias{fcwt_raw}
\title{Wraps CPP cwt function (computing a continuous wavelet transform)}
\usage{
fcwt_raw(input, startoctave, noctaves, nsuboctaves, sigma, nthreads, optplans)
}
\arguments{
\item{input}{A numeric vector representing the time series}

\item{startoctave}{Starting octave (corresponds to a scale of 2^startoctave)}

\item{noctaves}{Number of octaves to compute}

\item{nsuboctaves}{Computing steps per octave}

\item{sigma}{Parameter controlling time-frequency precision}

\item{nthreads}{Number of threads to use (FFTW)}

\item{optplans}{Use FFTW optimization plans}
}
\value{
Returns a numeric vector containing CWT information.
}
\description{
Wraps CPP cwt function (computing a continuous wavelet transform)
}
