% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{openmp_enabled}
\alias{openmp_enabled}
\title{Check for OpenMP support}
\usage{
openmp_enabled()
}
\value{
( \code{TRUE} | \code{FALSE} )
Returns \code{TRUE} if OpenMP support is enabled, \code{FALSE} otherwise.
}
\description{
This function checks if OpenMP support is enabled. It is responsible for
the multithreading capabilities of \code{\link[=fcwt]{fcwt()}}. If OpenMP is not enabled
the parameter \code{n_threads} of \code{\link[=fcwt]{fcwt()}} is ignored.
}
\details{
OpenMP can be used by the underlying fCWT library if
\itemize{
\item the fftw library on your platform was compiled with OpenMP support and
\item the fCWTr package itself was compiled with OpenMP support.
}

When using pre-built package binaries (like it is typical when you are using
R on Windows), whether or not these conditions are met, depend on the (CRAN)
build server. They cannot be influenced by the package author.

If the user is building the R package yourself, the user needs to make sure
that the fftw library on her platform are built with OpenMP support.
The fCWTr package is configured to use OpenMP if fftw-OpenMP support is
available.
}
