% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{print.fcwtr_scalogram}
\alias{print.fcwtr_scalogram}
\title{Print the scalogram}
\usage{
\method{print}{fcwtr_scalogram}(x, ...)
}
\arguments{
\item{x}{The "fcwtr_scalogram" object resulting from \code{\link[=fcwt]{fcwt()}} to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
\code{print()} prints its argument and returns it invisibly (via \code{invisible(x)}).
}
