% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma_freq_res_rel}
\alias{sigma_freq_res_rel}
\title{Calculates Relative Frequency Resolution}
\usage{
sigma_freq_res_rel(sigma)
}
\arguments{
\item{sigma}{The dimensionless positive parameter in \code{\link[=fcwt]{fcwt()}} controlling the wavelet spread.}
}
\value{
A number \code{p} representing the relative frequency resolution. I.e. at a given
resolution \code{f}, the frequency resolution is \code{f*p}.
}
\description{
For a given \code{sigma} this function calculates
the expected relative frequency resolution respectively.
}
\examples{
sigma_freq_res_rel(1)

}
\seealso{
Other sigma: 
\code{\link{sigma_from_freq_res}()},
\code{\link{sigma_from_freq_res_rel}()},
\code{\link{sigma_from_time_res}()},
\code{\link{sigma_res}()}
}
\concept{sigma}
