% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma_from_freq_res}
\alias{sigma_from_freq_res}
\title{Determine Sigma from a frequency resolution requirement}
\usage{
sigma_from_freq_res(res, freq)
}
\arguments{
\item{res}{The required frequency resolution in frequency units (see \code{\link[=u]{u()}}).}

\item{freq}{The reference frequency where the required resolution must hold.
Must be a quantity with frequency units (see \code{\link[=u]{u()}}).}
}
\value{
A dimensionless number controlling the FCWT wavelet spread,
corresponding to the \code{sigma} argument in \code{\link[=fcwt]{fcwt()}}.
}
\description{
Setting the correct value for \code{sigma} in \code{\link[=fcwt]{fcwt()}} for a particular
use case is important.
Sigma controls the wavelet spread used by the FCWT and so determines
the time-/frequency resolution (depending on a given reference frequency).
}
\details{
If you set a high frequency resolution, the time resolution
of the output signal will suffer, and vice versa. Determining the correct
balance for a particular use case can be crucial. This function determines
the value of sigma for a particular frequency resolution requirement.
}
\examples{
ref_freq <- u(440, "Hz")
sigma <- sigma_from_freq_res(u(7, "Hz"), ref_freq)

sigma_res(sigma, ref_freq)

}
\seealso{
Other sigma: 
\code{\link{sigma_freq_res_rel}()},
\code{\link{sigma_from_freq_res_rel}()},
\code{\link{sigma_from_time_res}()},
\code{\link{sigma_res}()}
}
\concept{sigma}
