% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{[.fcwtr_scalogram}
\alias{[.fcwtr_scalogram}
\title{Extract parts of a scalogram}
\usage{
\method{[}{fcwtr_scalogram}(x, i, j)
}
\arguments{
\item{x}{An "fcwtr_scalogram" object resulting from \code{\link[=fcwt]{fcwt()}}.}

\item{i, j}{Indices corresponding to time slices of the spectogram which specify
elements to extract. Indices are numeric vectors or empty (missing)
or NULL. Numeric values are coerced to integer or whole numbers as by as.integer
or for large values by \code{trunc()} (and hence truncated towards zero).

The time offset of the scalogram is adjusted to correspond to \code{min(i)}.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices
to select. Such vectors are recycled if necessary to match the corresponding extent.
i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.
When indexing arrays by [ a single argument i can be a matrix with as many
columns as there are dimensions of x; the result is then a vector with
elements corresponding to the sets of indices in each row of i.
An index value of NULL is treated as if it were integer(0).}
}
\value{
Another "fcwtr_scalogram" object that contains only part of the data.
}
\description{
Extract parts of a scalogram
}
\examples{
fcwt(
  sin((1:5000) * 2 * pi * 440 / 44100),
  sample_freq = 44100,
  n_freqs = 10
)[1:1000, 2:7]

}
