% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_series_sin}
\alias{time_series_sin}
\title{Superimposed sin waves of different frequencies and different amplitudes
(440 Hz, 880 Hz, 100 Hz, 500 Hz, 1200 Hz, 50 Hz, assuming a sample rate of 44100 Hz)}
\format{
## `time_series_sin`
A vector of float containing a signal time series.
}
\usage{
time_series_sin
}
\description{
Superimposed sin waves of different frequencies and different amplitudes
(440 Hz, 880 Hz, 100 Hz, 500 Hz, 1200 Hz, 50 Hz, assuming a sample rate of 44100 Hz)
}
\keyword{datasets}
