% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_piano_sample}
\alias{ts_piano_sample}
\title{A simple piano audio sample}
\format{
A numeric vector containing a time series signal.
}
\usage{
ts_piano_sample
}
\description{
Its sample frequency is 11025 Hz.
}
\keyword{datasets}
