% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_sin_440}
\alias{ts_sin_440}
\title{Pure sine wave}
\format{
A numeric vector containing a time series signal.
}
\usage{
ts_sin_440
}
\description{
Assuming a sample rate of 44100 Hz, the sine wave's frequency is 440 Hz.
}
\keyword{datasets}
