% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_sin_sin}
\alias{ts_sin_sin}
\title{A sinusoidal wave with varying frequency}
\format{
A numeric vector containing a time series signal.
}
\usage{
ts_sin_sin
}
\description{
The frequency itself is changing in a sinusoidal fashion.
}
\keyword{datasets}
