% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_sin_superpos}
\alias{ts_sin_superpos}
\title{Superimposed sine waves of different frequencies and different amplitudes}
\format{
A numeric vector containing a time series signal.
}
\usage{
ts_sin_superpos
}
\description{
Assuming a sample rate of 44100 Hz, the superimposed signals' frequencies are
440 Hz, 880 Hz, 100 Hz, 500 Hz, 1200 Hz and 50 Hz.
}
\keyword{datasets}
