% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{u}
\alias{u}
\title{Bestow units to a number}
\usage{
u(x, unit)
}
\arguments{
\item{x}{A numeric value to be bestowed with a unit.}

\item{unit}{A character vector indicating the unit. See \code{units::valid_udunits()} for
possible units.}
}
\description{
A wrapper around \code{units::set_unit()} mostly to have a short name to
create a dimensionful number.
}
\examples{
u(10, "km") # 10 kilometer
u(10, "mm") # 10 mm

# comparison taking into account the unit
u(100, "m") > u(1, "km")

}
\seealso{
\code{units::set_units()}
}
