% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{as.data.frame.fcwtr_scalogram}
\alias{as.data.frame.fcwtr_scalogram}
\title{Coerce the scalogram matrix to a data frame}
\usage{
\method{as.data.frame}{fcwtr_scalogram}(x, ...)
}
\arguments{
\item{x}{An object resulting from \code{\link[=fcwt]{fcwt()}}.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Internally, the scalogram resulting from \code{\link[=fcwt]{fcwt()}} is represented by
a numeric matrix. This method coerces this matrix into a reasonable
data frame. It is associated to a certain runtime cost.
}
