% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwt.R
\name{fcwt}
\alias{fcwt}
\title{Performs a fast continuous wavelet transform}
\usage{
fcwt(
  input,
  startoctave = 1L,
  noctave = 8L,
  nsuboctaves = 12L,
  sigma = 1,
  nthreads = 8,
  optplan = FALSE,
  abs = F
)
}
\arguments{
\item{input}{A numeric vector representing the time series}

\item{startoctave}{Starting octave (corresponds to a scale of 2^startoctave)}

\item{noctave}{The number of octaves to be processed.}

\item{nsuboctaves}{Computing steps per octave}

\item{sigma}{Parameter controlling time-frequency precision}

\item{nthreads}{Number of threads to use (FFTW)}

\item{optplan}{If TRUE uses the optimization plans that might have been
generated before by [create_optimization_schemes_raw].}

\item{abs}{Returns the absolute values instead of full
complex representations.}
}
\value{
Returns a numeric vector with dim = c(2, length(input), scales)
         where the first slot represents the real and imaginary part of the
         complex result, the second slot represents the time index, and
         the third slot represent different scales / frequencies.
}
\description{
This function call is a thin wrapper for the fcwt API.
}
