% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcwt_df.R
\name{plot.fcwt_df}
\alias{plot.fcwt_df}
\title{Plots a scalogram given a fcwt dataframe.}
\usage{
\method{plot}{fcwt_df}(x, ...)
}
\arguments{
\item{x}{A data frame as is returned by [fcwt.df].}
}
\value{
ggplot plot object
}
\description{
The x-axis corresponds to time, y-axis corresponds to frequency.
Note, the frequency axis is scaled logarithmically.
}
\examples{
# providing the sampling rate in addition to the time series is important to
# establishing a connection to physical units.
result_df <- fcwt_df(ts_sin_440,
  sampling_rate = 44100,
  time_resolution = 0.001
)
# plot the result
plot(result_df)

# now change the time-frequency uncertainty
result_df <- fcwt_df(ts_sin_440,
  sampling_rate = 44100,
  time_resolution = 0.001, sigma = 3
)
# plot the result again
plot(result_df)

}
