% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcwtr_scalogram.R
\name{rbind.fcwtr_scalogram}
\alias{rbind.fcwtr_scalogram}
\title{Combine scalograms in "time" direction}
\usage{
\method{rbind}{fcwtr_scalogram}(...)
}
\arguments{
\item{...}{One or more "fcwtr_scalogram" objects, generated by \code{fcwt()}.}
}
\value{
Returns a new time-wise combined "fcwtr_scalogram" object.
}
\description{
Given two or more scalograms with identical sampling frequencies, frequency
scales and sigma, it can be useful to combine several into a single object
creating a longer time series.
The function errs if these conditions are not satisfied.
}
\details{
The scalograms are stitched together in chronological fashion (i.e. the first
argument will the initial piece, etc.).
Time offset information is kept from the first piece.
}
\examples{
ts_sin_440 <- sin((1:5000) * 2 * pi * 440 / 44100)

res <-
  fcwt(
    ts_sin_440,
    sample_freq = u(44.1, "kHz"),
    freq_begin = u(50, "Hz"),
    freq_end = u(1000, "Hz"),
    n_freqs = 10,
    sigma = 5
  )

print(res)

# doubled scalogram
res_doubled <- rbind(res, res)

print(res_doubled)

}
