% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_sin_440}
\alias{ts_sin_440}
\title{Pure (long) sin wave of 440 Hz
assuming a sample rate of 44100 Hz)}
\format{
## `ts_sin_440`
A vector of float containing a signal time series.
}
\usage{
ts_sin_440
}
\description{
Pure (long) sin wave of 440 Hz
assuming a sample rate of 44100 Hz)
}
\keyword{datasets}
