% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ts_sin_sin}
\alias{ts_sin_sin}
\title{A sinusoidal wave with varying frequency.
The frequency itself is changing in a sinusoidal fashion.}
\format{
## `ts_sin_sin`
A vector of float containing a signal time series.
}
\usage{
ts_sin_sin
}
\description{
A sinusoidal wave with varying frequency.
The frequency itself is changing in a sinusoidal fashion.
}
\keyword{datasets}
