% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{u}
\alias{u}
\title{Augment physical unit}
\usage{
u(x, unit = "1")
}
\arguments{
\item{x}{A numeric vector to be augmented with a unit.}

\item{unit}{A character indicating the unit. See \code{\link[units:valid_udunits]{units::valid_udunits()}} and
\code{\link[units:valid_udunits]{units::valid_udunits_prefixes()}} for possible units.}
}
\description{
A wrapper around \code{\link[units:units]{units::set_units()}} mostly to have a short name to
create a dimensionful quantity.
}
\examples{
u(10, "km") # 10 kilometer
u(10, "mm") # 10 mm

# comparison taking into account the unit
u(100, "m") > u(1, "km")

}
\seealso{
\code{\link[units:units]{units::set_units()}}
}
